requirements.txt
scapy>=2.5.0
flask>=2.0.0
werkzeug>=2.0.0
python-dateutil>=2.8.2
gunicorn>=20.1.0
python-magic>=0.4.27
Flask-Limiter>=3.5.0